Develop role management functions

+++ backend/orbit-web-api/src/main/java/com/inmaytide/orbit/web/controller/sys/RoleController.java
@@ -1,12 +1,12 @@
 package com.inmaytide.orbit.web.controller.sys;
 
-import com.inmaytide.orbit.exceptions.InvalidParameterException;
+import com.inmaytide.orbit.exceptions.IllegalParameterException;
 import com.inmaytide.orbit.http.RestResponse;
 import com.inmaytide.orbit.model.basic.PageModel;
 import com.inmaytide.orbit.model.sys.Role;
 import com.inmaytide.orbit.service.sys.RoleService;
 import com.inmaytide.orbit.web.controller.BasicController;
-import org.hibernate.validator.constraints.NotBlank;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
@@ -38,14 +38,16 @@ public class RoleController extends BasicController {
     }
 
     @PostMapping("role")
-    public RestResponse add(@Validated Role role, BindingResult binding) {
+    public RestResponse add(@RequestBody @Validated Role role, BindingResult binding) {
         handleBindingResult(binding);
         return RestResponse.of(service.add(role));
     }
 
     @DeleteMapping("role")
-    public RestResponse remove(@NotBlank String ids, BindingResult binding) {
-        handleBindingResult(binding);
+    public RestResponse remove(String ids) {
+        if (StringUtils.isEmpty(ids)) {
+            throw new IllegalParameterException();
+        }
         service.remove(ids);
         return RestResponse.of(HttpStatus.OK);
     }
@@ -53,12 +55,12 @@ public class RoleController extends BasicController {
     @PutMapping("role")
     public RestResponse modify(@Validated Role role, BindingResult binding) {
         if (binding.hasErrors() || Objects.isNull(role.getId())) {
-            throw new InvalidParameterException();
+            throw new IllegalParameterException();
         }
         return RestResponse.of(service.modify(role));
     }
 
-    @GetMapping("checkCode/{code}/{id}")
+    @GetMapping("role/checkCode/{code}/{id}")
     public RestResponse checkCode(@PathVariable String code, @PathVariable Long id) {
         return RestResponse.of(service.checkCode(code, id));
     }
